moved initialisation out
